Added_Japanese_Translation-Text_Wrapping_for_CCJK
Added Japanese translation.

Also added a text wrapping function that will add newlines on a fixed character interval for CCJK writing, since the default text wrapping in Game Maker does not work with CCJK text.

Also changed the font for the main menu to Arial Unicode so that it displays all languages correctly.